+++ src/main/java/com/softwaremill/java_fp_example/contest/dmydlarz/jsoup/JsoupDocument.java
@@ -1,39 +1,43 @@
 package com.softwaremill.java_fp_example.contest.dmydlarz.jsoup;
 
-import org.jsoup.Jsoup;
+import org.jsoup.nodes.Attributes;
 import org.jsoup.nodes.Element;
+import org.jsoup.parser.Tag;
 
-import java.io.IOException;
-import java.net.MalformedURLException;
-import java.net.URL;
+public interface JsoupDocument {
+    Element head() throws Exception;
 
-class JsoupDocument implements Document {
-    private static final int DEFAULT_TIMEOUT = 10_000;
-    private final String uri;
-    private final int timeout;
+    final class Empty implements JsoupDocument {
+        @Override
+        public Element head() throws Exception {
+            return new Element("meta");
+        }
 
-    JsoupDocument(String uri, int timeout) {
-        this.uri = uri;
-        this.timeout = timeout;
-    }
-
-    JsoupDocument(String uri) {
-        this(uri, DEFAULT_TIMEOUT);
-    }
-
-    @Override
-    public Element head() throws Exception {
-        try {
-            return Jsoup.parse(new URL(uri), timeout).head();
-        } catch (MalformedURLException e) {
-            throw new Exception(String.format("Invalid uri passed: %s", uri), e);
-        } catch (IOException e) {
-            throw new Exception(String.format("Unable to connect to: %s", uri), e);
+        @Override
+        public String toString() {
+            return "empty document";
         }
     }
 
-    @Override
-    public String toString() {
-        return uri;
+    final class Fake implements JsoupDocument {
+        private static final String BASE_URI = "localhost";
+        private final String content;
+
+        Fake(String content) {
+            this.content = content;
+        }
+
+        @Override
+        public Element head() throws Exception {
+            Attributes attributes = new Attributes();
+            attributes.put("property", "og:image");
+            attributes.put("content", content);
+            return new Element(Tag.valueOf("meta"), BASE_URI, attributes);
+        }
+
+        @Override
+        public String toString() {
+            return String.format("fake document with og:image content: %s", content);
+        }
     }
 }
