+++ src/main/java/com/softwaremill/java_fp_example/contest/dmydlarz/jsoup/JsoupWebPage.java
@@ -5,17 +5,17 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 public final class JsoupWebPage implements WebPage {
-    private final Document document;
+    private final JsoupDocument jsoupDocument;
 
-    public JsoupWebPage(Document document) {
-        this.document = document;
+    public JsoupWebPage(JsoupDocument jsoupDocument) {
+        this.jsoupDocument = jsoupDocument;
     }
 
     public String facebookImage() throws Exception {
         try {
-            return new JsoupFacebookImage(document).uri();
+            return new JsoupFacebookImage(jsoupDocument).uri();
         } catch (Exception e) {
-            throw new Exception(String.format("Can't get facebook image for %s", document), e);
+            throw new Exception(String.format("Can't get facebook image for %s", jsoupDocument), e);
         }
     }
 }
